# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Тестовое задание.",
        "title": "Leadgen",
        "contact": {
            "name": "Константин",
            "url": "https://t.me/London68"
        },
        "version": "1.0"
    },
    "paths": {
        "/building/": {
            "post": {
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "summary": "Добавляет Строение, возвращает уникальный номер",
                "operationId": "building-handler-add",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название_записи",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Населенный пункт",
                        "name": "city",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Год",
                        "name": "year",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Этажность",
                        "name": "level",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.AddSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResp"
                        }
                    }
                }
            }
        },
        "/buildings/": {
            "get": {
                "produces": [
                    "application/json",
                    "multipart/form-data"
                ],
                "summary": "Получает Все объекты с базы. Есть фильтрация",
                "operationId": "building-handler-get",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название_записи",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Населенный пункт",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Год",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Этажность",
                        "name": "level",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Building.Building"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Building.Building": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "api.AddSuccess": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.ErrorResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}